requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
requests>=2.31.0
PyGithub>=2.1.1
python-dotenv>=1.0.0
markdown>=3.5.0
celery>=5.3.0
redis>=5.0.0
django-cors-headers>=4.3.0
gunicorn
psycopg2-binary
